riscv: cheri: Make various library routines cheri-ready
- Implement user_ptr.h for riscv-cheri
- Adapt vsprintf implementation to riscv-cheri and to a
  CHERI enabled kernel.
- Make various library routines ready for a CHERI enabled kernel.
  In particular this affects data structures that store additional
  bits in pointers.
- Handle all CHERI and non-CHERI cases in alignment macros for
  pointers ({USER_,}PTR_ALIGN{,_DOWN}).

+++ include/linux/assoc_array_priv.h
@@ -135,7 +135,7 @@ static inline void *assoc_array_ptr_to_leaf(const struct assoc_array_ptr *x)
 }
 
 static inline
-unsigned long __assoc_array_ptr_to_meta(const struct assoc_array_ptr *x)
+uintptr_t __assoc_array_ptr_to_meta(const struct assoc_array_ptr *x)
 {
 	return (uintptr_t)x &
 		~(ASSOC_ARRAY_PTR_SUBTYPE_MASK | ASSOC_ARRAY_PTR_TYPE_MASK);
